write comment and docstring
